fix extension mode
